+++ rustls/examples/internal/bench.rs
@@ -12,7 +12,7 @@ use std::sync::Arc;
 use std::time::{Duration, Instant};
 
 use rustls::client::{ClientSessionMemoryCache, NoClientSessionStorage};
-use rustls::crypto::Ring;
+use rustls::crypto::ring::Ring;
 use rustls::server::{
     AllowAnyAuthenticatedClient, NoClientAuth, NoServerSessionStorage, ServerSessionMemoryCache,
 };
